Tuesday Edits
Not much today.
Ive added some very basic code to check if a DB file exists and if it doesnt create it. This will hold the users message History so it can be replayed and dumped into the history box after closing the App.

Ive also made the validate IP address happen everytime a change is detected in the IPBox. You should no longer have to hit enter either.

I also added a hour minute timestamp to the message box.